% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp.data.2k100bp}
\alias{emp.data.2k100bp}
\title{EMP's 2k sample-subset with 100bp sequences.}
\format{
An object of class \code{dgCMatrix} with 161129 rows and 1856 columns.
}
\usage{
emp.data.2k100bp
}
\description{
A sparse matrix containing 100bp-sequence counts of the EMP's subset of 2,000 samples.
}
\keyword{datasets}
