% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp.data.2k150bp}
\alias{emp.data.2k150bp}
\title{EMP's 2k sample-subset with 150bp sequences.}
\format{
An object of class \code{dgCMatrix} with 91364 rows and 975 columns.
}
\usage{
emp.data.2k150bp
}
\description{
A sparse matrix containing 150bp-sequence counts of the EMP's subset of 2,000 samples.
}
\keyword{datasets}
