% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp.data.2k90bp}
\alias{emp.data.2k90bp}
\title{EMP's 2k sample-subset with 90bp sequences.}
\format{
An object of class \code{dgCMatrix} with 155002 rows and 2000 columns.
}
\usage{
emp.data.2k90bp
}
\description{
A sparse matrix containing 90bp-sequence counts of the EMP's subset of 2,000 samples.
}
\keyword{datasets}
