% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_sequences}
\alias{get_sequences}
\title{Get sequences as a SeqFastadna object}
\usage{
get_sequences(emp.obj, file.out = "emp_data.fasta", print.out = FALSE)
}
\arguments{
\item{emp.obj}{a dgCMatrix object containing otu-observation table}

\item{file.out}{Fasta file name to write}

\item{print.out}{Should a fasta file be printed instead ??}
}
\value{

}
\description{
Get sequences as a SeqFastadna object
}
\examples{

}
