% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs.metadata.2k90bp}
\alias{obs.metadata.2k90bp}
\title{EMP's 2k sample-subset; taxonomy of 90bp sequences.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 155002 rows and 8 columns.
}
\usage{
obs.metadata.2k90bp
}
\description{
A data table containing 90bp-sequences and the identified taxonomic levels.
}
\keyword{datasets}
