% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_metadata_150bp.R
\docType{data}
\name{observation_metadata_150bp}
\alias{observation_metadata_150bp}
\title{Observation taxonomy table}
\format{
Data frame with columns
\describe{
\item{sequence}{Observation-sequence found.}
\item{taxonomy1}{First taxonomic rank.}
\item{taxonomy2}{Second taxonomic rank.}
\item{taxonomy3}{Third taxonomic rank.}
\item{taxonomy4}{Fourth taxonomic rank.}
\item{taxonomy5}{Fifth taxonomic rank.}
\item{taxonomy6}{Sixth taxonomic rank.}
\item{taxonomy7}{Seventh taxonomic rank.}
}
}
\source{
\url{https://earthmicrobiome.org/}
}
\usage{
observation_metadata_150bp
}
\description{
This is a data set containing the related taxonomy
for the sequences originated from the Earth Microbiome Project.
}
\examples{
  head(observation_metadata_150bp)
}
\keyword{datasets}
