% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_metadata_sub2k150.R
\docType{data}
\name{observation_metadata_sub2k150}
\alias{observation_metadata_sub2k150}
\title{Observation taxonomy table}
\format{
Data frame with columns
\describe{
\item{sequence}{Observation-sequence found.}
\item{taxonomy1}{First taxonomic rank.}
\item{taxonomy2}{Second taxonomic rank.}
\item{taxonomy3}{Third taxonomic rank.}
\item{taxonomy4}{Fourth taxonomic rank.}
\item{taxonomy5}{Fifth taxonomic rank.}
\item{taxonomy6}{Sixth taxonomic rank.}
\item{taxonomy7}{Seventh taxonomic rank.}
}
}
\source{
\url{https://earthmicrobiome.org/}
}
\usage{
observation_metadata_sub2k150
}
\description{
This a 975-sample subset of the EMP (Earth Microbiome Project) taxonomy table containing 150bp-length sequences.
The samples were picked randomly and evenly across 17 habitats and then evenly across studies in each sample type.
}
\examples{
  str(observation_metadata_sub2k150)
}
\keyword{datasets}
