% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.metadata.sub2k}
\alias{sample.metadata.sub2k}
\title{EMP's 2k sample-subset; sample meta data.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2000 rows and 76 columns.
}
\usage{
sample.metadata.sub2k
}
\description{
A data table contaiing sample meta data for EMP's 2k-subset.
}
\keyword{datasets}
