% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\docType{data}
\name{sample_metadata}
\alias{sample_metadata}
\title{Sample metadata table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 975 rows and 76 columns.
}
\source{
\url{https://earthmicrobiome.org/}
}
\usage{
sample_metadata
}
\description{
This is a meta data file containing information
about samples originated from the Earth Microbiome Project.
}
\examples{
  head(sample_metadata)
}
\keyword{datasets}
