% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biom_data_150bp.R
\docType{data}
\name{biom_data_150bp}
\alias{biom_data_150bp}
\title{Biom data table}
\format{
Data frame with columns
\describe{
\item{sequence}{Observation-sequence found.}
\item{807.W.S.12.b}{Sample name.}
\item{2382.RU006.C181.RH.2.383.root.9.11.lane1.NoIndex.L001.sequences}{Sample name.}
\item{...}{Sample names.}
}
}
\source{
\url{https://earthmicrobiome.org/}
}
\usage{
biom_data_150bp
}
\description{
This a 976-sample subset of the EMP (Earth Microbiome Project) observation table containing 150bp-length sequences.
The samples were picked randomly and evenly across 17 habitats and then evenly across studies in each sample type.
}
\examples{
  head(biom_data_150bp)
}
\keyword{datasets}
